Misconfigured job
Notice the recurring declaration of `mapred.LocalJobRunner` in the job output. This report shows the job must have run on one member node of the cluster, but not in the cluster as a whole. As a result, there was no need to test for a delegation toke, which indicates the job ran in a Kerberized environment.